feat: add `begin` option for when to begin playing a video
This was specifically aimed towards youtube live streams, which contain
segments for the past 4 hours streamed. An additional option was added,
`liveBuffer` to work with `begin`.

The `refreshInterval` option was removed, the playlist will now only
refresh as the last segment in the playlist is approached. This works
better for streams played at a variable rate.

closes #7